+++ cdp/__init__.py
@@ -7,6 +7,7 @@ from cdp.cdp import Cdp
 from cdp.contract_invocation import ContractInvocation
 from cdp.faucet_transaction import FaucetTransaction
 from cdp.payload_signature import PayloadSignature
+from cdp.smart_contract import SmartContract
 from cdp.sponsored_send import SponsoredSend
 from cdp.trade import Trade
 from cdp.transaction import Transaction
@@ -32,4 +33,5 @@ __all__ = [
     "Trade",
     "SponsoredSend",
     "PayloadSignature",
+    "SmartContract",
 ]